mysite/requirements.txt
appdirs==1.4.4
astroid==1.6.6
backports.functools-lru-cache==1.6.1
colorama==0.4.3
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.1
djangorestframework==3.9.4
enum34==1.1.10
filelock==3.0.12
futures==3.3.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==4.3.21
lazy-object-proxy==1.5.1
mccabe==0.6.1
pathlib2==2.3.5
pylint==1.9.5
pyreadline==2.1
scandir==1.10.0
singledispatch==3.4.0.3
six==1.15.0
typing==3.7.4.3
virtualenv==20.0.29
wrapt==1.12.1
zipp==1.2.0